requirements/base.txt
argon2==0.1.10
argon2-cffi==18.1.0
Babel==2.6.0
backcall==0.1.0
boto3==1.9.18
botocore==1.12.18
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
decorator==4.3.0
Django==1.11
django-autoslug-1.9.4
django-cors-headers==2.4.0
django-extensions==2.1.0
django-filter==1.1.0
django-js-asset==1.1.0
django-mptt==0.9.1
django-phonenumber-field==2.0.0
django-storages==1.7.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docutils==0.14
gevent==1.3.6
google-api-python-client==1.7.4
google-auth==1.6.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.7
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.1
jmespath==0.9.3
Markdown==2.6.11
parso==0.3.1
pexpect==4.6.0
phonenumbers==8.9.10
phonenumberslite==8.9.14
pickleshare==0.7.4
Pillow==5.2.0
profanity==1.1
prompt-toolkit==1.0.15
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.4
pyotp==2.2.6
python-dateutil==2.7.3
python-slugify==1.2.5
pytz==2018.5
raven==6.9.0
requests==2.19.1
s3transfer==0.1.13
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
Unidecode==1.0.22
urllib3==1.23
wcwidth==0.1.7
